Split code into separate files
