+++ packages/app/src/components/difference-indicator/components/containers.tsx
@@ -1,0 +1,24 @@
+import css from '@styled-system/css';
+import styled from 'styled-components';
+
+export const IconContainer = styled.span(
+  css({
+    svg: {
+      mr: 1,
+      width: '19px',
+      height: '19px',
+    },
+  })
+);
+
+export const Container = styled.span(
+  css({
+    display: 'inline-block',
+    fontSize: 2,
+    svg: {
+      mr: 1,
+      width: '1.2em',
+      verticalAlign: 'text-bottom',
+    },
+  })
+);
