+++ packages/app/src/components/difference-indicator/components/sidebar-indicator.tsx
@@ -1,0 +1,41 @@
+import { DifferenceDecimal, DifferenceInteger } from '@corona-dashboard/common';
+import IconGelijk from '~/assets/gelijk.svg';
+import IconUp from '~/assets/pijl-omhoog.svg';
+import IconDown from '~/assets/pijl-omlaag.svg';
+import { Container, IconContainer } from './containers';
+
+export function SidebarIndicator({
+  value,
+}: {
+  value: DifferenceDecimal | DifferenceInteger;
+}) {
+  const { difference } = value;
+
+  if (difference > 0) {
+    return (
+      <Container>
+        <IconContainer color="red">
+          <IconUp />
+        </IconContainer>
+      </Container>
+    );
+  }
+
+  if (difference < 0) {
+    return (
+      <Container>
+        <IconContainer color="data.primary">
+          <IconDown />
+        </IconContainer>
+      </Container>
+    );
+  }
+
+  return (
+    <Container>
+      <IconContainer color="lightGray">
+        <IconGelijk />
+      </IconContainer>
+    </Container>
+  );
+}
