+++ packages/app/src/components/difference-indicator/index.ts
@@ -1,0 +1,2 @@
+export * from './difference-indicator';
+export * from './moving-average-difference-indicator';
